requirements.txt
Django==1.9.5
gunicorn==19.4
Markdown==2.6.6
psycopg2==2.6
Pillow==3.2.0
django-markupfield==1.4.0
django-tagging==0.4.3
django_compressor==2.0
django_cleanup==0.4.2